Add renderbuffer for performance, broke everything
This is an experiment. I want to go with chunks after this.

Adding materials significantly increases rendercalls and worsenes
performance dramatically.

I can fix it by giving the blocks only one material, and then use UV
coords for textures. However, this would mean that you cannot add custom
materials.

The next step would be to divide the world in chunks, see what blocks
are visible and which not, those which are not are not getting added in
the chunk mesh, the opposite for the others. Then do the same thing with
faces.